+++ wordpress/wp-content/plugins/minisite-manager/cookies.txt
@@ -1,0 +1,4 @@
+# Netscape HTTP Cookie File
+# https://curl.se/docs/http-cookies.html
+# This file was generated by libcurl! Edit at your own risk.
+
