+++ wordpress/wp-content/plugins/minisite-manager/src/Features/MinisiteListing/Commands/ListMinisitesCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Minisite\Features\MinisiteEditor\Commands;
+namespace Minisite\Features\MinisiteListing\Commands;
 
 /**
  * List Minisites Command
