+++ wordpress/wp-content/plugins/minisite-manager/src/Features/MinisiteListing/Http/ListingRequestHandler.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Minisite\Features\MinisiteListing\Http;
+
+use Minisite\Features\MinisiteListing\Commands\ListMinisitesCommand;
+
+/**
+ * Listing Request Handler
+ *
+ * SINGLE RESPONSIBILITY: Handle HTTP requests for listing functionality
+ * - Validates HTTP method
+ * - Extracts and sanitizes request data
+ * - Creates command objects
+ * - Handles nonce verification
+ */
+final class ListingRequestHandler
+{
+    /**
+     * Parse list minisites request
+     *
+     * @return ListMinisitesCommand|null
+     */
+    public function parseListMinisitesRequest(): ?ListMinisitesCommand
+    {
+        $currentUser = wp_get_current_user();
+        if (!$currentUser || !$currentUser->ID) {
+            return null;
+        }
+
+        // Get pagination parameters
+        $limit = isset($_GET['limit']) ? (int) $_GET['limit'] : 50;
+        $offset = isset($_GET['offset']) ? (int) $_GET['offset'] : 0;
+
+        // Validate limits
+        $limit = max(1, min(100, $limit)); // Between 1 and 100
+        $offset = max(0, $offset); // Non-negative
+
+        return new ListMinisitesCommand(
+            userId: $currentUser->ID,
+            limit: $limit,
+            offset: $offset
+        );
+    }
+}
