+++ wordpress/wp-content/plugins/minisite-manager/src/Features/MinisiteListing/Http/ListingResponseHandler.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Minisite\Features\MinisiteEditor\Http;
+namespace Minisite\Features\MinisiteListing\Http;
 
 /**
- * Editor Response Handler
+ * Listing Response Handler
  *
- * SINGLE RESPONSIBILITY: Handle HTTP responses and redirects
+ * SINGLE RESPONSIBILITY: Handle HTTP responses and redirects for listing functionality
  * - Manages redirects
  * - Sets HTTP headers
  * - Handles error responses
  */
-final class EditorResponseHandler
+final class ListingResponseHandler
 {
     /**
      * Redirect to login page
@@ -18,7 +18,6 @@ final class EditorResponseHandler
     public function redirectToLogin(?string $redirectTo = null): void
     {
         $loginUrl = home_url('/account/login');
-        
         if ($redirectTo) {
             $loginUrl .= '?redirect_to=' . urlencode($redirectTo);
         }
@@ -28,7 +27,7 @@ final class EditorResponseHandler
     }
 
     /**
-     * Redirect to sites page
+     * Redirect to sites listing page
      */
     public function redirectToSites(): void
     {
@@ -37,32 +36,11 @@ final class EditorResponseHandler
     }
 
     /**
-     * Redirect to specific URL
+     * Generic redirect
      */
     public function redirect(string $url): void
     {
         wp_redirect($url);
         exit;
-    }
-
-    /**
-     * Set 404 response
-     */
-    public function set404Response(): void
-    {
-        status_header(404);
-        nocache_headers();
-        
-        global $wp_query;
-        $wp_query->set_404();
-    }
-
-    /**
-     * Set 500 response
-     */
-    public function set500Response(): void
-    {
-        status_header(500);
-        nocache_headers();
     }
 }
